(WallpaperCropActivity) Remove redundant type casts

+++ src/ch/deletescape/wallpaperpicker/WallpaperCropActivity.java
@@ -368,7 +368,7 @@ public class WallpaperCropActivity extends Activity implements Handler.Callback 
         Point inSize = mCropView.getSourceDimensions();
 
         int cropRotation = mCropView.getImageRotation();
-        float cropScale = mCropView.getWidth() / (float) cropRect.width();
+        float cropScale = mCropView.getWidth() / cropRect.width();
 
         Matrix rotateMatrix = new Matrix();
         rotateMatrix.setRotate(cropRotation);
@@ -411,8 +411,8 @@ public class WallpaperCropActivity extends Activity implements Handler.Callback 
             cropRect.bottom += expandHeight;
         }
 
-        final int outWidth = (int) Math.round(cropRect.width() * cropScale);
-        final int outHeight = (int) Math.round(cropRect.height() * cropScale);
+        final int outWidth = Math.round(cropRect.width() * cropScale);
+        final int outHeight = Math.round(cropRect.height() * cropScale);
         CropAndFinishHandler onEndCrop = new CropAndFinishHandler(new Point(outWidth, outHeight),
                 shouldFadeOutOnFinish);
